Make min zoom, as calculated by `mz_calculate_poi_level`, a real number so that we can use it in the stylesheet to prioritise render order within zoom levels.

+++ data/apply-planet_osm_point.sql
@@ -10,7 +10,7 @@ CREATE INDEX planet_osm_point_place_index ON planet_osm_point(place) WHERE name 
 -- a partial index at some point in the future, depends on how the
 -- query performs.
 
-ALTER TABLE planet_osm_point ADD COLUMN mz_poi_min_zoom SMALLINT;
+ALTER TABLE planet_osm_point ADD COLUMN mz_poi_min_zoom REAL;
 -- the coalesce here is just an optimisation, as the poi level
 -- will always be NULL if all of the arguments are NULL.
 UPDATE planet_osm_point SET
